pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bullcapital-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "uvicorn>=0.35.0",
    "yfinance>=0.2.65",
    "requests==2.32.4",
    "beautifulsoup4==4.13.4",
    "html5lib>=1.1",
    "selenium>=4.34.1",
    "lxml>=6.0.0",
    "dotenv>=0.9.9",
    "pydantic[email]>=2.11.7",
    "python-multipart>=0.0.20",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "psycopg2>=2.9.10",
    "sqlalchemy>=2.0.41",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.2",
]

[tool.uv.workspace]
members = [
    "bullcapital-backend",
]
